BitJobs/bitjobs/requirements.txt
Django==1.10.6
whitenoise==3.2
psycopg2==2.6.0
hypothesis==3.7.0
fake-factory==0.5.2
django-bootstrap3==8.2.1
django-nose==1.4.4
nose==1.3.7
djangorestframework==3.6.2
django-webpack-loader==0.4.1
-e git://github.com/mplaton1/django-registration-rest-framework.git#egg=django-registration-rest-framework
django-static-precompiler==1.5
django-taggit==0.22.0
django-taggit-serializer==0.1.5
pytz==2016.10
requests==2.13.0
django-registration==2.2
django-fontawesome==0.3.1
django-money==0.10.2
gunicorn==19.6.0
dj_database_url==0.4.1
django_compressor==2.1
-e git://github.com/IOTeam42/django-cc#egg=django-cc